+++ ui/src/components/profile/QasProfile.vue
@@ -20,8 +20,8 @@
 </template>
 
 <script>
-import filterObject from '../../helpers/filter-object'
-import screen from '../../mixins/screen'
+import filterObjectMixin from '../../mixins/filter-object'
+import screenMixin from '../../mixins/screen'
 
 import QasAvatar from '../avatar/QasAvatar'
 import QasBox from '../box/QasBox'
@@ -34,7 +34,7 @@ export default {
     QasGridGenerator
   },
 
-  mixins: [screen],
+  mixins: [screenMixin, filterObjectMixin],
 
   props: {
     columns: {