+++ ui/src/mixins/index.js
@@ -3,11 +3,13 @@ import generatorMixin from './generator.js'
 import passwordMixin from './password.js'
 import screenMixin from './screen.js'
 import viewMixin from './view.js'
+import filterObjectMixin from './filter-object.js'
 
 export {
   contextMixin,
   generatorMixin,
   passwordMixin,
   screenMixin,
-  viewMixin
+  viewMixin,
+  filterObjectMixin
 }